implements delete profile picture docs

+++ src/controllers/profilePictureController.ts
@@ -97,7 +97,7 @@ const profilePictureController = {
                         res.status(404).json({ message: 'user not found' })
                     }
                 } else {
-                    res.status(403).json({ message: 'no access' })
+                    res.status(403).json({ message: 'could not access' })
                 }
             } else {
                 res.status(412).json({ message: 'missing id'})
@@ -141,40 +141,45 @@ const profilePictureController = {
 
     delete: async (req: CustomRequest, res: Response): Promise<void> => {
         try {
+            const { user } = req
             const { id } = req.query
 
             if (id) {
-                const picture = await prisma.profile_pictures.findUnique({
-                    where: { id }
-                })
-
-                if (picture) {
-                    await prisma.profile_pictures.delete({
+                if (user?.access_level === 'admin' || user?.id === id) {
+                    const picture = await prisma.profile_pictures.findUnique({
                         where: { id }
                     })
 
-                    res.status(200).json({ message: 'profile picture deleted' })
+                    if (picture) {
+                        await prisma.profile_pictures.delete({
+                            where: { id }
+                        })
 
-                    if (process.env.STORAGE_TYPE === 's3') {
-                        return s3
-                            .deleteObject({
-                                Bucket: process.env.BUCKET_NAME,
-                                Key: id
-                            })
-                            .promise()
-                            .then(response => {
-                                console.log(response.status)
-                            })
-                            .catch(response => {
-                                console.log(response.status)
-                            })
+                        res.status(200).json({ message: 'profile picture deleted' })
+
+                        if (process.env.STORAGE_TYPE === 's3') {
+                            return s3
+                                .deleteObject({
+                                    Bucket: process.env.BUCKET_NAME,
+                                    Key: id
+                                })
+                                .promise()
+                                .then(response => {
+                                    console.log(response.status)
+                                })
+                                .catch(response => {
+                                    console.log(response.status)
+                                })
+                        } else {
+                            return promisify(fs.unlink)(
+                                path.resolve(__dirname, '..', '..', 'tmp', 'uploads', picture.id)
+                            )
+                        }
                     } else {
-                        return promisify(fs.unlink)(
-                            path.resolve(__dirname, '..', '..', 'tmp', 'uploads', picture.id)
-                        )
+                        res.status(404).json({ message: 'profile picture not found' })
                     }
                 } else {
-                    res.status(404).json({ message: 'profile picture not found' })
+                    res.status(403).json({ message: 'could not access'})
                 }
             } else {
                 res.status(412).json({ message: 'missing id' })
